Configure and enable spec tests using Jasmine and Karma (triggered with Grunt).
Update metadata in package.json to reference the Awesomedoro app instead of the Bloc bootstrap project.

Change default port that Hapi uses to avoid collisions with other system services that run on :3000.

Replace Apache 2.0 license with ISC license to match what is described in the package.json file.

Connect Karma, Jasmine, and Grunt for test processing.